+++ vitest.setup.ts
@@ -1,4 +1,3 @@
-// Any setup scripts you might need go here
+import { config } from 'dotenv'
 
-// Load .env files
-import 'dotenv/config'
+config({ path: 'test.env', override: true })